requirements.txt
discord.py
python-dotenv
flask
flask-cors
mysql-connector-python
pytz